Auth: restructure store
