+++ frontend/src/modules/auth/auth-module.js
@@ -1,5 +1,5 @@
 import routes from '@/modules/auth/auth-routes'
-import store from '@/modules/auth/auth-store'
+import store from '@/modules/auth/store'
 
 export default {
   routes,
