+++ frontend/src/modules/auth/store/index.js
@@ -1,0 +1,13 @@
+import state from './state'
+import actions from './actions'
+import getters from './getters'
+import mutations from './mutations'
+
+export default {
+  namespaced: true,
+
+  state,
+  actions,
+  getters,
+  mutations
+}
