+++ frontend/src/modules/auth/store/state.js
@@ -1,0 +1,12 @@
+export default {
+  currentUser: null,
+  currentTenant: null,
+  loadingInit: true,
+  loadingEmailConfirmation: false,
+  loadingPasswordResetEmail: false,
+  loadingVerifyEmail: false,
+  loadingPasswordReset: false,
+  loadingPasswordChange: false,
+  loadingUpdateProfile: false,
+  loading: false
+}
