+++ frontend/src/premium/user/user-model.js
@@ -87,14 +87,12 @@ const fields = {
   }),
   password: new StringField('password', label('password'), {
     required: true,
-    min: 8
   }),
   passwordConfirmation: new StringField(
     'passwordConfirmation',
     label('passwordConfirmation'),
     {
       required: true,
-      min: 8
     }
   ),
   oldPassword: new StringField(
